requirements.txt
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
filelock==3.12.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
googlesearch-python==1.2.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
packaging==23.1
parsel==1.8.1
Protego==0.2.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rsa==4.9
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
zope.interface==6.0
